Update George_Simon_assign3
Overall it looks like you did a pretty good job on this assignment.  I like that you were creative with background colors, titles, button layout, etc.  However, I think it would be cool if you tried to add something like an image as a background.  That might make it look a little more interesting instead of just using different colors.  But if you have a theme you like I think that looks great.  I am curious about your camera functionality.  I used the same code as you did but I always receive an error that kills my app when I try to utilize it while testing. You may want to look at the book again at the chapter that we were supposed be following for this assignment.  Although your code functions well, it is not as efficient as it could be using the strategy this chapter in the book showed.  I completely understand this is all difficult and it is a hassle to try and learn different ways to do it but I just thought I would point it out as something to think about next time.  Overall, I would probably say your code ended up a little better than my own so nice work.